Breakout RectangleDrawable
