+++ src/Svg.Skia/Svg/RectangleDrawable.cs
@@ -1,0 +1,80 @@
+﻿#if !USE_PICTURE
+using SkiaSharp;
+using CropRect = SkiaSharp.SKImageFilter.CropRect;
+#endif
+#if USE_PICTURE
+using SKRect = Svg.Picture.Rect;
+#endif
+
+namespace Svg.Skia
+{
+    internal sealed class RectangleDrawable : DrawablePath
+    {
+        private RectangleDrawable()
+            : base()
+        {
+        }
+
+        public static RectangleDrawable Create(SvgRectangle svgRectangle, SKRect skOwnerBounds, DrawableBase? parent, Attributes ignoreAttributes = Attributes.None)
+        {
+            var drawable = new RectangleDrawable
+            {
+                Element = svgRectangle,
+                Parent = parent,
+                IgnoreAttributes = ignoreAttributes
+            };
+
+            drawable.IsDrawable = drawable.CanDraw(svgRectangle, drawable.IgnoreAttributes) && drawable.HasFeatures(svgRectangle, drawable.IgnoreAttributes);
+
+            if (!drawable.IsDrawable)
+            {
+                return drawable;
+            }
+
+            drawable.Path = svgRectangle.ToSKPath(svgRectangle.FillRule, skOwnerBounds, drawable.Disposable);
+            if (drawable.Path is null || drawable.Path.IsEmpty)
+            {
+                drawable.IsDrawable = false;
+                return drawable;
+            }
+
+            drawable.IsAntialias = SvgExtensions.IsAntialias(svgRectangle);
+
+            drawable.TransformedBounds = drawable.Path.Bounds;
+
+            drawable.Transform = SvgExtensions.ToSKMatrix(svgRectangle.Transforms);
+
+            bool canDrawFill = true;
+            bool canDrawStroke = true;
+
+            if (SvgExtensions.IsValidFill(svgRectangle))
+            {
+                drawable.Fill = SvgExtensions.GetFillSKPaint(svgRectangle, drawable.TransformedBounds, ignoreAttributes, drawable.Disposable);
+                if (drawable.Fill is null)
+                {
+                    canDrawFill = false;
+                }
+            }
+
+            if (SvgExtensions.IsValidStroke(svgRectangle, drawable.TransformedBounds))
+            {
+                drawable.Stroke = SvgExtensions.GetStrokeSKPaint(svgRectangle, drawable.TransformedBounds, ignoreAttributes, drawable.Disposable);
+                if (drawable.Stroke is null)
+                {
+                    canDrawStroke = false;
+                }
+            }
+
+            if (canDrawFill && !canDrawStroke)
+            {
+                drawable.IsDrawable = false;
+                return drawable;
+            }
+
+            // TODO: Transform _skBounds using _skMatrix.
+            drawable.TransformedBounds = drawable.Transform.MapRect(drawable.TransformedBounds);
+
+            return drawable;
+        }
+    }
+}
